requirements.txt
# ============================================================================
# ImgCraft - Production Dependencies for Render Deployment
# ============================================================================

# Core Flask Framework
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# WSGI Server for Production
gunicorn==21.2.0

# Image Processing Libraries
Pillow==10.1.0
opencv-python-headless==4.8.1.78
scikit-image==0.22.0
numpy==1.24.3
piexif==1.1.3

# AI Background Removal
rembg>=2.0.50
onnxruntime==1.16.3

# Authentication & Database
supabase==2.3.0
python-dotenv==1.0.0
pyjwt==2.8.0
postgrest==0.13.0
realtime==1.0.0
storage3==0.7.0
supafunc==0.3.0

# Payment Processing
razorpay==1.4.1

# HTTP Requests & Utilities
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# Date/Time Utilities
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0

# Logging & Monitoring (Optional but recommended)
colorlog==6.8.0

# Security
cryptography==41.0.7

# ============================================================================
# Production Server Configuration
# ============================================================================
# Run with: gunicorn --bind 0.0.0.0:$PORT app:app
# Render will automatically set the PORT environment variable
